+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/block/BaseCropBlock.java
@@ -42,6 +42,10 @@ public abstract class BaseCropBlock extends CropBlock {
         return ItemStack.EMPTY;
     }
 
+    protected int getAgeAfterKnife() {
+        return this.getMaxAge() - 1;
+    }
+
     @SuppressWarnings("deprecation")
     @Override
     public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand interactionHand, @NotNull BlockHitResult blockHitResult) {
@@ -53,7 +57,7 @@ public abstract class BaseCropBlock extends CropBlock {
             if (matureAge) {
                 popResource(level, pos, this.getBaseItemStack());
                 level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
-                BlockState newBlockState = this.getStateForAge(age - 2);
+                BlockState newBlockState = this.getStateForAge(getAgeAfterKnife());
                 level.setBlock(pos, newBlockState, 2);
                 level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, newBlockState));
                 stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));