+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/block/CottonCropBlock.java
@@ -13,65 +13,32 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.CropBlock;
-import net.minecraft.world.level.block.FarmBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraft.world.level.gameevent.GameEvent;
-import net.minecraft.world.level.storage.loot.LootParams;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.List;
-
 public class CottonCropBlock extends BaseCropBlock {
 
-    public static final IntegerProperty AGE = IntegerProperty.create("age", 0, 5);
+    public static final IntegerProperty AGE = IntegerProperty.create("age", 0, 6);
     private static final VoxelShape[] SHAPE_BY_AGE = new VoxelShape[] {
             Block.box(0.0, 0.0, 0.0, 16.0, 5.0, 16.0),
             Block.box(0.0, 0.0, 0.0, 16.0, 7.0, 16.0),
             Block.box(0.0, 0.0, 0.0, 16.0, 9.0, 16.0),
             Block.box(0.0, 0.0, 0.0, 16.0, 11.0, 16.0),
             Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0),
-            Block.box(0.0, 0.0, 0.0, 16.0, 15.0, 16.0)
+            Block.box(0.0, 0.0, 0.0, 16.0, 15.0, 16.0),
+            Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0)
     };
 
     public CottonCropBlock(Properties properties) {
         super(properties);
-        this.registerDefaultState(this.stateDefinition.any().setValue(this.getAgeProperty(), 0));
-    }
-
-    @Override
-    public @NotNull VoxelShape getShape(@NotNull BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos pos, @NotNull CollisionContext context) {
-        return SHAPE_BY_AGE[this.getAge(state)];
-    }
-
-    @Override
-    public int getMaxAge() {
-        return 5;
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public @NotNull List<ItemStack> getDrops(@NotNull BlockState state, LootParams.@NotNull Builder builder) {
-        List<ItemStack> dropsOriginal = super.getDrops(state, builder);
-        int randomCotton = 2 + builder.getLevel().random.nextInt(4);
-        int randomCottonSeeds = 2 + builder.getLevel().random.nextInt(2);
-        if (!dropsOriginal.isEmpty()) {
-            return dropsOriginal;
-        }
-        return state.getValue(AGE) == 5 ?  List.of(new ItemStack(ItemRegistry.COTTON.get(), randomCotton), new ItemStack(this, randomCottonSeeds)) : List.of(new ItemStack(this, 1));
-    }
-
-    @Override
-    protected @NotNull ItemLike getBaseSeedId() {
-        return ItemRegistry.COTTON_SEEDS.get();
     }
 
     @Override
@@ -81,35 +48,55 @@ public class CottonCropBlock extends BaseCropBlock {
     }
 
     @Override
+    public int getMaxAge() {
+        return 5;
+    }
+
+    @Override
+    protected @NotNull ItemLike getBaseSeedId() {
+        return ItemRegistry.COTTON_SEEDS.get();
+    }
+
+    @Override
+    protected ItemStack getBaseItemStack() {
+        return ItemRegistry.COTTON.get().getDefaultInstance();
+    }
+
+    @Override
+    protected int getAgeAfterKnife() {
+        return 2;
+    }
+
+    @Override
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         builder.add(AGE);
     }
 
     @Override
-    public boolean isValidBonemealTarget(@NotNull LevelReader levelReader, @NotNull BlockPos pos, @NotNull BlockState state, boolean $$3) {
-        return state.getValue(AGE) < 4;
+    public @NotNull VoxelShape getShape(@NotNull BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos pos, @NotNull CollisionContext context) {
+        return SHAPE_BY_AGE[this.getAge(state)];
     }
 
-    @SuppressWarnings("deprecation")
     @Override
     public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand interactionHand, @NotNull BlockHitResult blockHitResult) {
-        int age = state.getValue(AGE);
-        boolean reachedCottonAge = age == 5;
-        if (!reachedCottonAge && player.getItemInHand(interactionHand).is(Items.BONE_MEAL)) {
-            return InteractionResult.PASS;
-        } else if (reachedCottonAge) {
-            ItemStack stack = player.getItemInHand(interactionHand);
-            if (stack.getItem() instanceof KnifeItem) {
-                int randomCotton = 2 + level.random.nextInt(4);
-                int randomCottonSeeds = 2 + level.random.nextInt(2);
-                popResource(level, pos, new ItemStack(ItemRegistry.COTTON.get(), randomCotton));
-                popResource(level, pos, new ItemStack(ItemRegistry.COTTON_SEEDS.get(), randomCottonSeeds));
+        int age = this.getAge(state);
+        boolean matureAge = age == this.getMaxAge();
+        boolean deadAge = age == this.getMaxAge() + 1;
+        ItemStack stack = player.getItemInHand(interactionHand);
+        if (stack.getItem() instanceof KnifeItem) {
+            if (matureAge) {
+                popResource(level, pos, this.getBaseItemStack());
                 level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
-                BlockState newBlockState = state.setValue(AGE, 4);
+                BlockState newBlockState = this.getStateForAge(age - 2);
                 level.setBlock(pos, newBlockState, 2);
                 level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, newBlockState));
                 stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));
-                return InteractionResult.sidedSuccess(level.isClientSide);
+            } else if (deadAge) {
+                int random = level.random.nextInt(2);
+                popResource(level, pos, new ItemStack(random == 4 ? this.getBaseSeedId() : Items.STICK, level.random.nextInt(2)));
+                level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
+                level.setBlock(pos, Blocks.AIR.defaultBlockState(), 2);
+                stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));
             }
         }
         return super.use(state, level, pos, player, interactionHand, blockHitResult);